1. SplashActivity 생성 2. IntroActivity 생성 및 파이어베이스 관련 코드 이동
