+++ .idea/misc.xml
@@ -3,7 +3,9 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/gjaischool-184/AndroidStudioProjects/EV_CHECK/app/src/main/res/layout/activity_intro.xml" value="0.33559782608695654" />
         <entry key="..\:/Users/gjaischool-184/AndroidStudioProjects/EV_CHECK/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/gjaischool-184/AndroidStudioProjects/EV_CHECK/app/src/main/res/layout/activity_splash.xml" value="0.33559782608695654" />
       </map>
     </option>
   </component>
