+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,22 @@
         android:theme="@style/Theme.EV_CHECK"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".auth.IntroActivity"
+            android:exported="false" />
+        <activity
+            android:name=".SplashActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+        <intent-filter>
+            <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
+            <category android:name="android.intent.category.LAUNCHER" />
+        </intent-filter>
+        </activity>
+
+        <activity
+            android:name=".MainActivity"
+            android:exported="false">
+
         </activity>
     </application>
