+++ app/src/main/java/com/example/ev_check/SplashActivity.kt
@@ -1,0 +1,22 @@
+package com.example.ev_check
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.os.Handler
+import com.example.ev_check.auth.IntroActivity
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+
+        // SplashActivity에서 IntroActivity로 3초 후 이동
+        // 3초지연 --> SubThread --> **Handler --> MainThread
+        Handler().postDelayed({
+            val intent = Intent(this, IntroActivity::class.java)
+            startActivity(intent)
+            finish()
+        }, 3000)
+    }
+}
