+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,10 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
     <style name="Theme.EV_CHECK" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+
+        <!-- 상단바 제거 -->
+        <item name="windowNoTitle">true</item>
+
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/purple_500</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
